Upgrade to eleventy 1.0 🎉 (#111)
## Trivial

`s/slug/slugify`

## Changes

- More robust test triggering (Now gets triggered by test changes or by the build actually finishing)
- Fixed bad tests for draft posts.

## Not so great

Changed the `addHash` filter to read from source directories instead of `_site`. This works but it would not work if a file would be synthesized into the output dir. This is necessary due to a race condition that only happens in `watch` mode where `_site` doesn't exist during re-build.